# Docs/kindohub-api-spec.yaml
openapi: 3.0.3
info:
  title: KindoHub API
  version: 1.0.0
  description: API REST para la gestion integral de una asociacion de padres y madres (APA/AMPA). Incluye gestion de familias, alumnos, cursos, anotaciones, usuarios, formas de pago, estados de asociado y sistema de auditoria con logs.
  contact:
    name: KindoHub Development Team
    email: dev@kindohub.com
  license:
    name: Proprietary
    url: https://kindohub.com/license
servers:
  - url: https://api.kindohub.com
    description: Servidor de produccion
  - url: https://staging-api.kindohub.com
    description: Servidor de staging
  - url: http://localhost:5000
    description: Desarrollo local
...
paths:
...
  /api/usuarios/{username}:
    get:
      tags:
        - Usuarios
      summary: Obtener usuario por nombre
      operationId: getUserByUsername
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDto'
        '404':
          description: Usuario no encontrado
        '401':
          description: No autenticado
  /api/usuarios:
    get:
      tags:
        - Usuarios
      summary: Listar todos los usuarios
      operationId: getAllUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioDto'
  /api/usuarios/register:
    post:
      tags:
        - Usuarios
      summary: Registrar nuevo usuario
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrarUsuarioDto'
      responses:
        '200':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDto'
        '400':
          description: Datos invalidos
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UsuarioDto:
      type: object
      properties:
        Id:
          type: integer
        Username:
          type: string
        IsAdmin:
          type: boolean
        IsActive:
          type: boolean
        Rol:
          type: string
    RegistrarUsuarioDto:
      type: object
      required:
        - Username
        - Password
        - ConfirmPassword
      properties:
        Username:
          type: string
        Password:
          type: string
          format: password
        ConfirmPassword:
          type: string
          format: password
...
security:
  - BearerAuth: []